store: move database backends to libstore
