+++ po/POTFILES.in
@@ -83,14 +83,8 @@ gui/mutt_window.c
 gui/reflow.c
 gui/terminal.c
 handler.c
-hcache/bdb.c
-hcache/gdbm.c
 hcache/hcache.c
-hcache/kc.c
-hcache/lmdb.c
-hcache/qdbm.c
 hcache/serialize.c
-hcache/tc.c
 hdrline.c
 help.c
 history/history.c
@@ -205,6 +199,14 @@ smtp.c
 sort.c
 state.c
 status.c
+store/bdb.c
+store/gdbm.c
+store/kc.c
+store/lmdb.c
+store/qdbm.c
+store/store.c
+store/tc.c
+store/tdb.c
 system.c
 version.c
 wcscasecmp.c
