+++ store/bdb.c
@@ -39,7 +39,7 @@
 #include <sys/types.h>
 #include <unistd.h>
 #include "mutt/lib.h"
-#include "backend.h"
+#include "lib.h"
 #include "globals.h"
 
 /**
