+++ store/qdbm.c
@@ -35,7 +35,7 @@
 #include <stdbool.h>
 #include <villa.h>
 #include "mutt/lib.h"
-#include "backend.h"
+#include "lib.h"
 #include "globals.h"
 
 /**
