+++ store/tc.c
@@ -34,7 +34,7 @@
 #include <tcbdb.h>
 #include <tcutil.h>
 #include "mutt/lib.h"
-#include "backend.h"
+#include "lib.h"
 #include "globals.h"
 
 /**
